handle serverless template
